Remove all binaries
