--- src/Client/Client/.vs/Client/v17/DocumentLayout.json
@@ -1,37 +1,0 @@
-{
-  "Version": 1,
-  "WorkspaceRootPath": "D:\\4course\\APKS\\Labs\\csad2425ki45vorobeyaa02\\src\\Client\\Client\\",
-  "Documents": [
-    {
-      "AbsoluteMoniker": "D:0:0:{7A2724FF-77C8-4E7C-86DE-C38C1BC3EA21}|Client\\Client.vcxproj|D:\\4course\\APKS\\Labs\\csad2425ki45vorobeyaa02\\src\\Client\\Client\\Client\\main.cpp||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}",
-      "RelativeMoniker": "D:0:0:{7A2724FF-77C8-4E7C-86DE-C38C1BC3EA21}|Client\\Client.vcxproj|solutionrelative:Client\\main.cpp||{D0E1A5C6-B359-4E41-9B60-3365922C2A22}"
-    }
-  ],
-  "DocumentGroupContainers": [
-    {
-      "Orientation": 0,
-      "VerticalTabListWidth": 256,
-      "DocumentGroups": [
-        {
-          "DockedWidth": 200,
-          "SelectedChildIndex": 0,
-          "Children": [
-            {
-              "$type": "Document",
-              "DocumentIndex": 0,
-              "Title": "main.cpp",
-              "DocumentMoniker": "D:\\4course\\APKS\\Labs\\csad2425ki45vorobeyaa02\\src\\Client\\Client\\Client\\main.cpp",
-              "RelativeDocumentMoniker": "Client\\main.cpp",
-              "ToolTip": "D:\\4course\\APKS\\Labs\\csad2425ki45vorobeyaa02\\src\\Client\\Client\\Client\\main.cpp",
-              "RelativeToolTip": "Client\\main.cpp",
-              "ViewState": "AQIAAB4AAAAAAAAAAAAwwEYAAAAAAAAA",
-              "Icon": "ae27a6b0-e345-4288-96df-5eaf394ee369.000677|",
-              "WhenOpened": "2024-10-29T13:31:57.593Z",
-              "EditorCaption": ""
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}
