--- src/Client/Client/Client.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.9.34607.119
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Client", "Client\Client.vcxproj", "{7A2724FF-77C8-4E7C-86DE-C38C1BC3EA21}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7A2724FF-77C8-4E7C-86DE-C38C1BC3EA21}.Debug|x64.ActiveCfg = Debug|x64
-		{7A2724FF-77C8-4E7C-86DE-C38C1BC3EA21}.Debug|x64.Build.0 = Debug|x64
-		{7A2724FF-77C8-4E7C-86DE-C38C1BC3EA21}.Debug|x86.ActiveCfg = Debug|Win32
-		{7A2724FF-77C8-4E7C-86DE-C38C1BC3EA21}.Debug|x86.Build.0 = Debug|Win32
-		{7A2724FF-77C8-4E7C-86DE-C38C1BC3EA21}.Release|x64.ActiveCfg = Release|x64
-		{7A2724FF-77C8-4E7C-86DE-C38C1BC3EA21}.Release|x64.Build.0 = Release|x64
-		{7A2724FF-77C8-4E7C-86DE-C38C1BC3EA21}.Release|x86.ActiveCfg = Release|Win32
-		{7A2724FF-77C8-4E7C-86DE-C38C1BC3EA21}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {29E27854-02B2-4080-8852-DD9B3EAE6693}
-	EndGlobalSection
-EndGlobal
