--- src/Client/Client/Client/x64/Debug/Client.exe.recipe
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project>
-  <ProjectOutputs>
-    <ProjectOutput>
-      <FullPath>D:\4course\APKS\Labs\csad2425ki45vorobeyaa02\src\Client\Client\x64\Debug\Client.exe</FullPath>
-    </ProjectOutput>
-  </ProjectOutputs>
-  <ContentFiles />
-  <SatelliteDlls />
-  <NonRecipeFileRefs />
-</Project>
