--- src/Client/Client/Client/x64/Debug/Client.log
@@ -1,3 +1,0 @@
-﻿  main.cpp
-D:\4course\APKS\Labs\csad2425ki45vorobeyaa02\src\Client\Client\Client\main.cpp(7,58): warning C4267: 'argument': conversion from 'size_t' to 'DWORD', possible loss of data
-  Client.vcxproj -> D:\4course\APKS\Labs\csad2425ki45vorobeyaa02\src\Client\Client\x64\Debug\Client.exe
