--- src/Client/Client/Client/x64/Debug/Client.tlog/Cl.items.tlog
@@ -1,1 +1,0 @@
-D:\4course\APKS\Labs\csad2425ki45vorobeyaa02\src\Client\Client\Client\main.cpp;D:\4course\APKS\Labs\csad2425ki45vorobeyaa02\src\Client\Client\Client\x64\Debug\main.obj
