--- src/Client/Client/Client/x64/Debug/Client.tlog/Client.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.36.32532:TargetPlatformVersion=10.0.22621.0:
-Debug|x64|D:\4course\APKS\Labs\csad2425ki45vorobeyaa02\src\Client\Client\|
