--- src/Client/Client/Client/x64/Debug/Client.tlog/link.secondary.1.tlog
@@ -1,2 +1,0 @@
-^D:\4COURSE\APKS\LABS\CSAD2425KI45VOROBEYAA02\SRC\CLIENT\CLIENT\CLIENT\X64\DEBUG\MAIN.OBJ
-D:\4course\APKS\Labs\csad2425ki45vorobeyaa02\src\Client\Client\Client\x64\Debug\Client.ilk
